+++ src/pages/shop/work-orders/[id].tsx
@@ -1,3 +1,5 @@
+import WorkOrderEditor from "@components/workOrders/WorkOrderEditor";
+import WorkOrderMetadataDialog from "@components/workOrders/WorkOrderMetadataDialog";
 import { UserType } from "@prisma/client";
 import {
   getWorkOrderByIdActionBuilder,
@@ -10,15 +12,10 @@ import { GetServerSideProps, NextPage } from "next";
 import Head from "next/head";
 import { useRouter } from "next/router";
 import { Button } from "primereact/button";
-import { Dialog } from "primereact/dialog";
-import { Dropdown } from "primereact/dropdown";
-import { Editor } from "primereact/editor";
-import { InputText } from "primereact/inputtext";
 import { TabPanel, TabView } from "primereact/tabview";
 import { Toast } from "primereact/toast";
 import React, { useEffect, useRef, useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
-import { WorkOrderStatus } from "src/types/workOrder";
 
 const WorkOrder: NextPage = () => {
   const router = useRouter();
@@ -221,201 +218,11 @@ const WorkOrderPage: React.FC<{}> = () => {
           <p>Last Saved: {formatDate(new Date(workOrder.updateTime))}</p>
         )}
       </div>
-      <MetadataDialog
+      <WorkOrderMetadataDialog
         isVisible={isEditMetaDataDialogVisible}
         onHide={handleHideEditMetaDataDialog}
       />
     </div>
-  );
-};
-
-const WorkOrderEditor: React.FC<{
-  body: string;
-  updateBody: (body: string) => void;
-}> = (props) => {
-  const { body, updateBody } = props;
-
-  const editorToolbar = () => {
-    return (
-      <div>
-        <span className="ql-formats">
-          <select className="ql-header" defaultValue="0">
-            <option value="1">Heading</option>
-            <option value="2">Subheading</option>
-            <option value="0">Normal</option>
-          </select>
-          <select className="ql-font">
-            <option></option>
-            <option value="serif"></option>
-            <option value="monospace"></option>
-          </select>
-        </span>
-        <span className="ql-formats">
-          <button className="ql-bold" aria-label="Bold"></button>
-          <button className="ql-italic" aria-label="Italic"></button>
-          <button className="ql-underline" aria-label="Underline"></button>
-        </span>
-        <span className="ql-formats">
-          <select className="ql-color"></select>
-          <select className="ql-background"></select>
-        </span>
-        <span className="ql-formats">
-          <button
-            type="button"
-            className="ql-list"
-            value="ordered"
-            aria-label="Ordered List"
-          ></button>
-          <button
-            type="button"
-            className="ql-list"
-            value="bullet"
-            aria-label="Unordered List"
-          ></button>
-        </span>
-        <span className="ql-formats">
-          <button
-            type="button"
-            className="ql-clean"
-            aria-label="Remove Styles"
-          ></button>
-        </span>
-      </div>
-    );
-  };
-
-  return (
-    <Editor
-      // Moving this to CSS file breaks it
-      style={{ height: "50vh" }}
-      value={body}
-      onTextChange={(e) => updateBody(e.htmlValue ?? "")}
-      headerTemplate={editorToolbar()}
-    />
-  );
-};
-
-const MetadataDialog: React.FC<{
-  isVisible: boolean;
-  onHide: () => void;
-}> = (props) => {
-  const { query } = useRouter();
-  const { id } = query;
-
-  const dispatch = useDispatch();
-  const workOrder = useSelector(WorkOrderSelectors.getWorkOrder);
-
-  const [workOrderStatus, setWorkOrderStatus] = useState<
-    WorkOrderStatus | undefined
-  >(workOrder?.status);
-  const [workOrderTitle, setWorkOrderTitle] = useState<string | undefined>(
-    workOrder?.title
-  );
-  const [workOrderAssignedEmployeeEmail, setWorkOrderAssignedEmployeeEmail] =
-    useState<string | undefined>(workOrder?.employee?.email);
-
-  useEffect(() => {
-    if (workOrder) {
-      setWorkOrderStatus(workOrder.status);
-      setWorkOrderTitle(workOrder.title);
-      if (workOrder.employee) {
-        setWorkOrderAssignedEmployeeEmail(workOrder.employee.email);
-      }
-    }
-  }, [workOrder]);
-
-  const handleSave = () => {
-    if (typeof id === "string") {
-      dispatch(
-        patchWorkOrderByIdActionBuilder(id, {
-          title: workOrderTitle,
-          employee_email: workOrderAssignedEmployeeEmail,
-          status: workOrderStatus,
-        })
-      );
-    }
-
-    props.onHide();
-  };
-
-  const handleWorkOrderStatusUpdate = (status: string) => {
-    if (status === "Pending") {
-      setWorkOrderStatus(WorkOrderStatus.PENDING);
-    } else if (status === "In Progress") {
-      setWorkOrderStatus(WorkOrderStatus.IN_PROGRESS);
-    } else if (status === "Completed") {
-      setWorkOrderStatus(WorkOrderStatus.COMPLETED);
-    }
-  };
-
-  const footer = () => {
-    return (
-      <div>
-        <Button
-          className={`p-button-success ${styles.workOrderMetadataDialogFooterContainerButtons}`}
-          icon="pi pi-save"
-          label="Save"
-          aria-label="Save"
-          onClick={handleSave}
-        />
-        <Button
-          className={`p-button-danger ${styles.workOrderMetadataDialogFooterContainerButtons}`}
-          icon="pi pi-times"
-          label="Cancel"
-          aria-label="Cancel"
-          onClick={props.onHide}
-        />
-      </div>
-    );
-  };
-
-  return (
-    <Dialog
-      visible={props.isVisible}
-      onHide={props.onHide}
-      footer={footer}
-      header="Edit Metadata"
-    >
-      <div>
-        <label htmlFor="workOrderTitle">Title</label>
-        <br />
-        <InputText
-          id="workOrderTitle"
-          name="title"
-          placeholder={workOrderTitle}
-          value={workOrderTitle}
-          onChange={(e) => setWorkOrderTitle(e.target.value)}
-          className={styles.workOrderMetadataDialogTitleTextBox}
-        />
-        <br />
-
-        <label htmlFor="workOrderAssignedEmployee">
-          Assigned Employee (Email)
-        </label>
-        <br />
-        <InputText
-          id="workOrderAssignedEmployee"
-          name="assignedEmployee"
-          placeholder={workOrderAssignedEmployeeEmail}
-          value={workOrderAssignedEmployeeEmail}
-          onChange={(e) => setWorkOrderAssignedEmployeeEmail(e.target.value)}
-          className={styles.workOrderMetadataDialogEmployeeEmailTextBox}
-        />
-        <br />
-
-        <label htmlFor="workOrderStatus">Status</label>
-        <br />
-        <Dropdown
-          id="workOrderStatus"
-          name="status"
-          options={["Pending", "In Progress", "Completed"]}
-          placeholder={workOrderStatus}
-          value={workOrderStatus}
-          onChange={(e) => handleWorkOrderStatusUpdate(e.target.value)}
-          className={styles.workOrderMetadataDialogStatusSelector}
-        />
-      </div>
-    </Dialog>
   );
 };
