+++ src/styles/pages/WorkOrders.module.css
@@ -33,27 +33,6 @@
   width: 10rem;
 }
 
-.workOrderMetadataDialogFooterContainerButtons {
-  width: 10rem !important;
-}
-
-.workOrderMetadataDialogTitleTextBox {
-  width: 100%;
-  margin-top: 1rem;
-  margin-bottom: 1rem;
-}
-
-.workOrderMetadataDialogEmployeeEmailTextBox {
-  width: 100%;
-  margin-top: 1rem;
-  margin-bottom: 1rem;
-}
-
-.workOrderMetadataDialogStatusSelector {
-  width: 100%;
-  margin-top: 1rem;
-}
-
 @media (max-width: 600px) {
   .workOrderPageContainer {
     margin-left: 0rem;
